Fix fire-and-forget crash-time emergency protection + comprehensive async/decimal audit fixes (#345)
* Initial plan

* Fix fire-and-forget crash-time emergency protection - block until completion


* Update documentation to reflect emergency protection fix


* Fix async patterns in S6_S11_Bridge - replace GetAwaiter().GetResult() with Task.Wait()


* Partial decimal migration for Zones - constants, interfaces, ZoneSnapshot to decimal


* Complete decimal migration for Zones - all ATR and price calculations now use decimal


* Eliminate double casts in TradingSystemIntegrationService - RLAgent.MarketData now uses decimal


---------

+++ src/Zones/ZoneContracts.cs
@@ -7,7 +7,7 @@ public sealed record Zone(
     ZoneSide Side,
     decimal PriceLow,
     decimal PriceHigh,
-    double Pressure,     // 0..1 (touches, confluence, recency)
+    decimal Pressure,     // 0..1 (touches, confluence, recency)
     int TouchCount,
     DateTime LastTouchedUtc,
     ZoneState State)
@@ -22,10 +22,10 @@ public sealed record Zone(
 public sealed record ZoneSnapshot(
     Zone? NearestDemand,
     Zone? NearestSupply,
-    double DistToDemandAtr,   // >=0
-    double DistToSupplyAtr,   // >=0
-    double BreakoutScore,     // 0..1 about NEAREST OPPOSING zone
-    double ZonePressure,      // 0..1 of NEAREST OPPOSING zone
+    decimal DistToDemandAtr,   // >=0
+    decimal DistToSupplyAtr,   // >=0
+    decimal BreakoutScore,     // 0..1 about NEAREST OPPOSING zone
+    decimal ZonePressure,      // 0..1 of NEAREST OPPOSING zone
     DateTime Utc);
 
 public interface IZoneService
@@ -38,5 +38,5 @@ public interface IZoneService
 public interface IZoneFeatureSource
 {
     // Provides zone features to the brain/feature-bus each bar
-    (double distToDemandAtr, double distToSupplyAtr, double breakoutScore, double zonePressure) GetFeatures(string symbol);
+    (decimal distToDemandAtr, decimal distToSupplyAtr, decimal breakoutScore, decimal zonePressure) GetFeatures(string symbol);
 }
